feat: show filters inside dropdown buttons on search page
